--- .github/workflows/template_change_test.yml
@@ -1,17 +1,0 @@
-name: Rename Template
-
-on: pull_request
-
-jobs:
-  rename-template:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v1
-
-      - name: Set Up JDK
-        uses: actions/setup-java@v1
-        with:
-          java-version: 1.11
-
-      - name: Rename & Build
-        run: ./gradlew renameTemplate build
