+++ app/src/main/java/com/adammcneilly/rustybeer/theme/Color.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MagicNumber")
 
-package template.theme
+package com.adammcneilly.rustybeer.theme
 
 import androidx.compose.ui.graphics.Color
