+++ build.gradle
@@ -33,7 +33,6 @@ buildscript {
 
 apply from: "buildscripts/versions.gradle"
 apply from: "buildscripts/githooks.gradle"
-apply from: "buildscripts/setup.gradle"
 
 subprojects {
     apply from: "../buildscripts/ktlint.gradle"
